backend/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
pgvector>=0.2.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# LightRAG Core & Transitive Dependencies
lightrag-hku[api]>=1.0.0
nano-vectordb>=0.0.4
json_repair>=0.25.0
networkx>=3.0
pypinyin>=0.50.0
pipmaster>=0.5.0
xlsxwriter>=3.1.0
configparser>=6.0.0
tenacity>=8.2.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.4.0
Pillow>=10.0.0

# LLM Providers & Support
openai>=1.12.0
ollama>=0.1.0
aioboto3>=12.0.0
boto3>=1.34.0
aiohttp>=3.9.0
google-api-core>=2.0.0
google-genai>=0.1.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.0
pdf2image>=1.17.0

# Machine Learning / Embeddings
transformers>=4.37.0
torch>=2.0.0
sentence-transformers>=2.3.0
